Make sure embeddings don't fail (#120)
Make sure embeddings don't fail with rate limits

- When documents are ready to be embedded, they are now added to an EmbeddingBacklog
- The EmbeddingTaskExecutor takes document from the EmbeddingBacklog and tries to embed them.
- If this fails, it tries again after some time.

I also tested other methods to prevent hitting the rate limit. When we are using an embedding model with a larger context window, we may increase the batch size for requests and not the length of the actual documents. Thus, the chunks of test are not too large, and we still get a performance boost due to fewer calls to the model.

* Set embedded additional property when listing sources

* Handling of errors and correct processing of tasks if errors occur

* Update JavaDoc of the EmbeddingBacklog interface to better reflect requirements for implementations

* Remove EmbeddingTasks from EmbeddingBacklog if the source is deleted

When a source is deleted, the EmbeddingBacklog is tasked to remove the associated task, if it exists.

* Adding priorities to the EmbeddingBacklog

* Move EmbeddingBacklog related classes to backlog package

* Make defer atomic in InMemoryEmbeddingBacklog

* Wait for completion of task if it was in the backlog and currently processed

* Downgrade spring ai bom to 1.0.0 due to swagger ui not working

* Fixing return value of remove method